project delete confirmation alert

+++ src/main/java/dk/easv/gui/controller/InspectProjectController.java
@@ -1,15 +1,20 @@
 package dk.easv.gui.controller;
 
 import dk.easv.be.Project;
+import dk.easv.bll.logic.ProjectDisplay;
+import dk.easv.bll.util.PopupUtil;
 import dk.easv.gui.model.ProjectModel;
 import dk.easv.gui.util.BlurEffectUtil;
 import javafx.fxml.FXML;
 import javafx.fxml.Initializable;
+import javafx.scene.Scene;
 import javafx.scene.control.Button;
+import javafx.scene.control.ButtonType;
 import javafx.scene.layout.BorderPane;
 import javafx.stage.Stage;
 
 import java.net.URL;
+import java.util.Optional;
 import java.util.ResourceBundle;
 
 public class InspectProjectController implements Initializable {
@@ -35,10 +40,16 @@ public class InspectProjectController implements Initializable {
     }
 
     public void deleteProject(){
-        try {
-            projectModel.deleteProject(project);
-        } catch (Exception e) {
-            throw new RuntimeException(e);
+        Optional<ButtonType> result = PopupUtil.showConfirmationAlert("Confirm deletion", "Are you sure you want to delete this project?");
+        if (result.isPresent() && result.get() == ButtonType.OK) {
+            try {
+                projectModel.deleteProject(project);
+            } catch (Exception e) {
+                throw new RuntimeException(e);
+            }
+            Stage stage = (Stage) deleteButton.getScene().getWindow();
+            stage.close();
+            BlurEffectUtil.removeBlurEffect(borderPane);
         }
     }
